+++ pex/http.py
@@ -212,7 +212,7 @@ class RequestsContext(Context):
   def open(self, link):
     # requests does not support file:// -- so we must short-circuit manually
     if link.local:
-      return open(link.path, 'rb')  # noqa: T802
+      return open(link.local_path, 'rb')  # noqa: T802
     for attempt in range(self._max_retries + 1):
       try:
         return StreamFilelike(self._session.get(
@@ -248,7 +248,7 @@ if requests:
 class CachedRequestsContext(RequestsContext):
   """A requests-based Context with CacheControl support."""
 
-  DEFAULT_CACHE = '~/.pex/cache'
+  DEFAULT_CACHE = os.path.join(ENV.PEX_ROOT, 'cache')
 
   def __init__(self, cache=None, **kw):
     self._cache = os.path.realpath(os.path.expanduser(cache or self.DEFAULT_CACHE))