+++ pex/package.py
@@ -3,7 +3,7 @@
 
 import os
 
-from pkg_resources import EGG_NAME, parse_version, safe_name
+from pkg_resources import EGG_NAME, parse_version, safe_name, safe_version
 
 from .archiver import Archiver
 from .base import maybe_requirement
@@ -133,7 +133,7 @@ class SourcePackage(Package):
 
   @property
   def raw_version(self):
-    return safe_name(self._raw_version)
+    return safe_version(self._raw_version)
 
   # SourcePackages are always compatible as they can be translated to a distribution.
   def compatible(self, identity, platform=Platform.current()):
@@ -167,7 +167,7 @@ class EggPackage(Package):
 
   @property
   def raw_version(self):
-    return safe_name(self._raw_version)
+    return safe_version(self._raw_version)
 
   @property
   def py_version(self):
@@ -223,6 +223,16 @@ class WheelPackage(Package):
       if tag in self._supported_tags:
         return True
     return False
+
+  def __eq__(self, other):
+    return (self._name == other._name and
+            self._raw_version == other._raw_version and
+            self._py_tag == other._py_tag and
+            self._abi_tag == other._abi_tag and
+            self._arch_tag == other._arch_tag)
+
+  def __hash__(self):
+    return hash((self._name, self._raw_version, self._py_tag, self._abi_tag, self._arch_tag))
 
 
 Package.register(SourcePackage)