+++ tests/test_pex_builder.py
@@ -6,10 +6,11 @@ import stat
 import zipfile
 from contextlib import closing
 
+import pytest
 from twitter.common.contextutil import temporary_dir
 from twitter.common.dirutil import safe_mkdir
 
-from pex.compatibility import nested
+from pex.compatibility import WINDOWS, nested
 from pex.pex import PEX
 from pex.pex_builder import PEXBuilder
 from pex.testing import write_simple_pex as write_pex
@@ -57,15 +58,19 @@ def test_pex_builder():
 
 
 def test_pex_builder_shebang():
-  pb = PEXBuilder()
-  pb.set_shebang('foobar')
+  def builder(shebang):
+    pb = PEXBuilder()
+    pb.set_shebang(shebang)
+    return pb
 
-  with temporary_dir() as td:
-    target = os.path.join(td, 'foo.pex')
-    pb.build(target)
-    expected_preamble = b'#!foobar\n'
-    with open(target, 'rb') as fp:
-      assert fp.read(len(expected_preamble)) == expected_preamble
+  for pb in builder('foobar'), builder('#!foobar'):
+    for b in pb, pb.clone():
+      with temporary_dir() as td:
+        target = os.path.join(td, 'foo.pex')
+        b.build(target)
+        expected_preamble = b'#!foobar\n'
+        with open(target, 'rb') as fp:
+          assert fp.read(len(expected_preamble)) == expected_preamble
 
 
 def test_pex_builder_compilation():
@@ -102,6 +107,7 @@ def test_pex_builder_compilation():
     build_and_check(td3, True)
 
 
+@pytest.mark.skipif(WINDOWS, reason='No hardlinks on windows')
 def test_pex_builder_copy_or_link():
   with nested(temporary_dir(), temporary_dir(), temporary_dir()) as (td1, td2, td3):
     src = os.path.join(td1, 'exe.py')
@@ -112,13 +118,17 @@ def test_pex_builder_copy_or_link():
       pb = PEXBuilder(path, copy=copy)
       pb.add_source(src, 'exe.py')
 
-      s1 = os.stat(src)
-      s2 = os.stat(os.path.join(path, 'exe.py'))
-      is_link = (s1[stat.ST_INO], s1[stat.ST_DEV]) == (s2[stat.ST_INO], s2[stat.ST_DEV])
-      if copy:
-        assert not is_link
-      else:
-        assert is_link
+      path_clone = os.path.join(path, '__clone')
+      pb.clone(into=path_clone)
+
+      for root in path, path_clone:
+        s1 = os.stat(src)
+        s2 = os.stat(os.path.join(root, 'exe.py'))
+        is_link = (s1[stat.ST_INO], s1[stat.ST_DEV]) == (s2[stat.ST_INO], s2[stat.ST_DEV])
+        if copy:
+          assert not is_link
+        else:
+          assert is_link
 
     build_and_check(td2, False)
     build_and_check(td3, True)